% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypr.R
\name{keypr}
\alias{keypr}
\title{keypr main() Wrapper}
\usage{
keypr(...)
}
\arguments{
\item{...}{Arguments to \code{keypr_main()}.}
}
\value{
Quits R with an error code: 0 if no errors, > 0 otherwise.
}
\description{
A try-catch wrapper around \code{keypr_main()}.
}
\seealso{
\code{\link[=keypr_main]{keypr_main()}}.
}
