% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypr-package.R
\docType{package}
\name{keypr-package}
\alias{keypr-package}
\alias{_PACKAGE}
\title{keypr: Credentials Management Using 'GPG' Encryption, 'YAML' Files and 'CLI'}
\description{
A 'CLI' and a set of corresponding functions to work with an 
    encrypted credentials storage.  A credential is a login-password pair, each 
    linked to a "service" identified by a unique name.  This information is 
    serialized to 'YAML' and stored encrypted with a symmetric cipher using 
    'GPG'.  'keypr' implements a convenient 'CLI' to work with the storage 
    (querying / adding new records).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x1o/keypr}
  \item Report bugs at \url{https://github.com/x1o/keypr/issues}
}

}
\keyword{internal}
